requirements.txt
discordpy
discord Interaction
yt_dlp
openai
json
FlightRadar24API
asyncio
